+++ hutool-script/src/main/java/com/xiaoleilu/hutool/script/JavaScriptEngine.java
@@ -7,7 +7,6 @@ import javax.script.Compilable;
 import javax.script.CompiledScript;
 import javax.script.Invocable;
 import javax.script.ScriptContext;
-import javax.script.ScriptEngine;
 import javax.script.ScriptEngineFactory;
 import javax.script.ScriptEngineManager;
 import javax.script.ScriptException;
@@ -18,8 +17,6 @@ import javax.script.ScriptException;
  *
  */
 public class JavaScriptEngine extends FullSupportScriptEngine{
-	
-	ScriptEngine engine;
 	
 	public JavaScriptEngine() {
 		super(new ScriptEngineManager().getEngineByName("javascript"));