+++ hutool-script/src/test/java/com/xiaoleilu/hutool/script/ScriptUtilTest.java
@@ -1,0 +1,17 @@
+package com.xiaoleilu.hutool.script;
+
+import org.junit.Test;
+
+/**
+ * 脚本单元测试类
+ * @author looly
+ *
+ */
+public class ScriptUtilTest {
+	
+	@Test
+	public void compileTest() {
+		ScriptUtil.eval("println('Script test!');");
+	}
+	
+}