+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### 新特性
 * MailUtil邮件工具类支持附件
+* Convert增加int、long、short与bytes之间的转换
+* BeetlUtil增加更多简化方法
+* extra模块中模板相关工具类移入template包中
+* ScriptUtil增加eval方法，执行脚本快捷方法
 
 ### Bug修复
 * 修复FileUtil.normalize导致的路径修复问题